adding various separations
