+++ .env.example
@@ -1,15 +1,6 @@
 # ---------- DCX ----------
-PORT=3000
-EXTERNAL_PORT=3000
-EXTERNAL_HOSTNAME=localhost
-SERVICE_NAME='decentralized credential exchange'
 SERVICE_ID=dcx
-
-# ---------- DHT ----------
-
-# Replace with a list of your DWN endpoints
-# Defaults to TBD DHT gateway
-DHT_GATEWAY_ENDPOINT=https://dht.example.com
+SERVICE_NAME='decentralized credential exchange'
 
 # ---------- Web5 ----------
 # Replace with a comma delimited list of your DWN endpoints
